+++ Demo/src/main/java/com/braintreepayments/demo/PayPalMessagingFragment.kt
@@ -5,6 +5,7 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.LinearLayout
+import com.braintreepayments.api.PayPalMessagingColor
 import com.braintreepayments.api.PayPalMessagingListener
 import com.braintreepayments.api.PayPalMessagingLogoType
 import com.braintreepayments.api.PayPalMessagingOfferType
@@ -26,12 +27,12 @@ class PayPalMessagingFragment : BaseFragment(), PayPalMessagingListener {
 
         val payPalMessagingRequest = PayPalMessagingRequest(
             amount = 2.0,
-            placement = null,
+            pageType = null,
             offerType = PayPalMessagingOfferType.PAY_LATER_LONG_TERM,
             buyerCountry = "US",
             logoType = PayPalMessagingLogoType.PRIMARY,
             textAlignment = PayPalMessagingTextAlignment.CENTER,
-            color = null
+            color = PayPalMessagingColor.BLACK
         )
 
         val payPalMessagingView = PayPalMessagingView(braintreeClient, requireActivity())