+++ PayPalMessaging/src/main/java/com/braintreepayments/api/PayPalMessagingOfferType.kt
@@ -28,7 +28,7 @@ enum class PayPalMessagingOfferType {
      */
     PAYPAL_CREDIT_NO_INTEREST;
 
-    internal val offerTypeRawValue: PayPalMessageOfferType
+    internal val internalValue: PayPalMessageOfferType
         get() = when (this) {
             PAY_LATER_SHORT_TERM -> PayPalMessageOfferType.PAY_LATER_SHORT_TERM
             PAY_LATER_LONG_TERM -> PayPalMessageOfferType.PAY_LATER_LONG_TERM