+++ PayPalMessaging/src/test/java/com/braintreepayments/api/PayPalMessagingOfferTypeUnitTest.kt
@@ -9,7 +9,7 @@ class PayPalMessagingOfferTypeUnitTest {
     fun `test pay later short term offer type returns raw value pay later short term`() {
         assertEquals(
             PayPalMessageOfferType.PAY_LATER_SHORT_TERM,
-            PayPalMessagingOfferType.PAY_LATER_SHORT_TERM.offerTypeRawValue
+            PayPalMessagingOfferType.PAY_LATER_SHORT_TERM.internalValue
         )
     }
 
@@ -17,7 +17,7 @@ class PayPalMessagingOfferTypeUnitTest {
     fun `test pay later long term offer type returns raw value pay later long term`() {
         assertEquals(
             PayPalMessageOfferType.PAY_LATER_LONG_TERM,
-            PayPalMessagingOfferType.PAY_LATER_LONG_TERM.offerTypeRawValue
+            PayPalMessagingOfferType.PAY_LATER_LONG_TERM.internalValue
         )
     }
 
@@ -25,7 +25,7 @@ class PayPalMessagingOfferTypeUnitTest {
     fun `test pay later pay in one offer type returns raw value pay later pay in 1`() {
         assertEquals(
             PayPalMessageOfferType.PAY_LATER_PAY_IN_1,
-            PayPalMessagingOfferType.PAY_LATER_PAY_IN_ONE.offerTypeRawValue
+            PayPalMessagingOfferType.PAY_LATER_PAY_IN_ONE.internalValue
         )
     }
 
@@ -33,7 +33,7 @@ class PayPalMessagingOfferTypeUnitTest {
     fun `test PayPal credit no interest offer type returns raw value PayPal credit no interest`() {
         assertEquals(
             PayPalMessageOfferType.PAYPAL_CREDIT_NO_INTEREST,
-            PayPalMessagingOfferType.PAYPAL_CREDIT_NO_INTEREST.offerTypeRawValue
+            PayPalMessagingOfferType.PAYPAL_CREDIT_NO_INTEREST.internalValue
         )
     }
 }