+++ PayPalMessaging/src/test/java/com/braintreepayments/api/PayPalMessagingViewUnitTest.kt
@@ -3,8 +3,10 @@ package com.braintreepayments.api
 import android.content.Context
 import androidx.test.core.app.ApplicationProvider
 import com.braintreepayments.api.Configuration.Companion.fromJson
+import com.paypal.messages.config.message.PayPalMessageConfig
 import io.mockk.verify
 import io.mockk.mockk
+import io.mockk.mockkObject
 import io.mockk.slot
 import junit.framework.TestCase.assertEquals
 import org.junit.Before
@@ -63,6 +65,7 @@ class PayPalMessagingViewUnitTest {
     @Test
     fun `test start with valid configuration calls onPayPalMessagingLoading delegate and sends analytics`() {
         val payPalConfiguration: Configuration = fromJson(Fixtures.CONFIGURATION_WITH_LIVE_PAYPAL)
+        mockkObject(PayPalMessageConfig)
         val braintreeClient = MockkBraintreeClientBuilder()
             .configurationSuccess(payPalConfiguration)
             .build()
@@ -73,5 +76,9 @@ class PayPalMessagingViewUnitTest {
 
         verify { listener.onPayPalMessagingLoading() }
         verify { braintreeClient.sendAnalyticsEvent("paypal-messaging:create-view:started") }
+        verify { PayPalMessageConfig.setGlobalAnalytics(
+                integrationName = "BT_SDK",
+                integrationVersion = BuildConfig.VERSION_NAME
+        ) }
     }
 }